pyproject.toml
[project]
name = "transient"
version = "0.1.0"
description = "Transient.AI Ramp-Up Program - Financial AI Agents"
authors = [{name = "Transient AI"}]
license = {text = "Proprietary"}
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "financial-datasets",
    "pandas>=2.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.scripts]
desk-agent-service = "src.service.main:main"

[tool.uv]
dev-dependencies = [
    "pytest>=9.0.2",
    "build>=1.0.3",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
